+++ release_notes.md
@@ -7,6 +7,7 @@
 - Added Dutch (Netherlands) language
 - Added Slovak (Slovakia) language
 - Added new configuration option `ask_for_further_commands` to control Alexa's behavior for asking further commands.
+- Start a conversation with prompt from Home Assistant (thanks to [t07que])
 
 ---
 
@@ -19,3 +20,4 @@
 - Adicionado o idioma Holandês (Holanda)
 - Adicionado o idioma Eslovaco (Eslováquia)
 - Adicionada nova opção de configuração `ask_for_further_commands` para controlar o comportamento da Alexa ao perguntar por mais comandos.
+- Adicionado opção para iniciar uma conversa com Alexa a partir de um prompt do Home Assistant